\name{statgdx-package}
\alias{statgdx-package}
\alias{statgdx}
\docType{package}
\title{Exchanging data between GAMS and R}
\description{
This package implements an interface between GAMS and R.  It includes
functions to transfer data between GDX (the GAMS data format) and R, a
function to call GAMS from within R, query functions for the meta-data
in GDX, and other related utilities.
}
\details{
  This package is really a hack.
}
\author{Steve Dirkse.  Maintainer: \email{R@gams.com}}
\keyword{ package }
\keyword{ data }
\keyword{ optimize }
\keyword{ interface }
% \examples{ }
